23.02.2023

+++ Entities/Concrete/Car.cs
@@ -15,7 +15,7 @@ namespace Entities
         public int ColorId { get; set; }
         public string ModelName { get; set; }
         public int ModelYear { get; set; }
-        public decimal DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public string Description { get; set; }
 
     }
